+++ service-worker.js
@@ -9,6 +9,7 @@ const ASSETS = [
   './css/modules.css',
   './css/tokens.css',
   './js/core/app.js',
+  './js/platform/shell/themeController.js',
   './js/core/centralStore.js',
   './js/core/domUpdate.js',
   './js/core/eventDelegation.js',